+++ cartoframes/context/__init__.py
@@ -1,0 +1,9 @@
+from .api_context import APIContext
+
+
+def create_context(credentials, session=None):
+    return APIContext(credentials, session)
+
+
+__all__ = [
+]