+++ test/viz/helpers/test_size_continuous_layer.py
@@ -36,6 +36,7 @@ class TestSizeContinuousLayerHelper(unittest.TestCase):
         self.assertEqual(layer.legend._type['line'], 'size-continuous-line')
         self.assertEqual(layer.legend._title, 'name')
         self.assertEqual(layer.legend._description, '')
+        self.assertEqual(layer.legend._footer, '')
 
     def test_size_continuous_layer_point(self):
         "should create a point type layer"